requirements.txt
Flask>=2.2.2
SQLAlchemy>=1.4.46
python-dotenv>=0.21.0